makes all scene items into prefabs, so removes the dependency on the TestScene unity file
