+++ Assets/JuliusCaesarBust.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 4d1d7d32f89d7184f823f908349a081a
+folderAsset: yes
+timeCreated: 1511453982
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
