+++ Assets/JuliusCaesarBust/FBX.meta
@@ -1,0 +1,5 @@
+fileFormatVersion: 2
+guid: d04753157dd1cb249a4fc2fde13e483d
+folderAsset: yes
+DefaultImporter:
+  userData:
