+++ Assets/JuliusCaesarBust/Textures.meta
@@ -1,0 +1,5 @@
+fileFormatVersion: 2
+guid: b4427824481ebab45bf90cdc30800f0a
+folderAsset: yes
+DefaultImporter:
+  userData:
