+++ Assets/JuliusCaesarBust/readme_first.txt
@@ -1,0 +1,32 @@
+First of all, please take your time to read the documentation of the asset you downloaded and play a little bit with it. 
+After that, please come back here and keep reading. It just will take you a few seconds.
+
+
+
+If you like my work and wonder how to support me a little bit more, here are some options for you that would help me a lot to continue working:
+
+
+
+1- Rate and comment the asset at the asset store.
+
+
+2- Buying any other of my paid assets:
+
+		https://www.assetstore.unity3d.com/en/#!/search/page=1/sortby=popularity/query=publisher:6052
+
+
+3- Donations (PayPal): snicolast81@gmail.com
+
+
+4- Hiring me, if you are a company that looks for a 3D artist.
+
+
+
+
+Thank you very much for your contribution and help, 
+
+
+Kind regards,
+
+
+Sergi
