+++ Assets/JuliusCaesarBust/readme_first.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f54ff4d4b1503194d87582f7a19104c3
+timeCreated: 1451419751
+licenseType: Store
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
