+++ Assets/New Terrain.asset.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d118e15e6a33ff94cb59aa8d0e90d56f
+timeCreated: 1511450725
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
