+++ Assets/Project/Prefabs/Cowboy.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 7f2d8e6f9c6aa4c658f91af7db8144f7
+timeCreated: 1511535133
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
