+++ Assets/Project/Prefabs/DialogueParent.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ebeb9de8034ca48f484d720d79b18a3c
+timeCreated: 1511535139
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
