+++ Assets/Project/Prefabs/ShoppingListParent.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 29c3abcc8ae1c47c69978d5b34239245
+timeCreated: 1511535143
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
