+++ Assets/Project/Scripts/Vive/PlayerController.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: fc6bd5b6d2612fe4db90a7a4913b21e3
+timeCreated: 1511389027
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
