--- Assets/Project/Scripts/ViveController.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ffd74be8313b6fb4da8d4455d4ecbdee
-folderAsset: yes
-timeCreated: 1510846797
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
