--- Assets/Project/Scripts/ViveController/WandController.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 6ac5afd6ff06c9349a89c31207277cd0
-timeCreated: 1510846812
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
