+++ Assets/Wooden_table_and_chair.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 750a67d7e5a229c43b4fad30537f7801
+folderAsset: yes
+timeCreated: 1511449089
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
