+++ Assets/Wooden_table_and_chair/FBX.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 93b8bf0b6f7bc7f41aca647aad12634d
+folderAsset: yes
+timeCreated: 1511449089
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
