+++ Assets/Wooden_table_and_chair/FBX/Materials.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: f00c03b95acf1ce4abd2484751fb5c26
+folderAsset: yes
+timeCreated: 1511449089
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
