+++ Assets/Wooden_table_and_chair/FBX/Materials/table_chair_01.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 658d0a6601bdfc54e9bffb7089fcbe11
+timeCreated: 18446744011573954816
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
